Add files via upload
This project contains the tutorial of retrieving data from the input box and print them on a table on the same page where the table row will be increased dynamically. To learn this you need to know the basics of Html, CSS & JavaScript.
In this project , we'll be handling the basics of data processing and display with Javascript. The inputs your user gives in the input text boxes should be processed by our Javascript code and displayed in the desired format in a table.